# (WIP) Optimize String#gsub with block