--- a/src/http/common.cr
+++ b/src/http/common.cr
@@ -298,14 +298,13 @@ module HTTP
   def self.quote_string(string, io)
     # Escaping rules: https://evolvis.org/pipermail/evolvis-platfrm-discuss/2014-November/000675.html
 
-    string.each_byte do |byte|
-      case byte
-      when '\t'.ord, ' '.ord, '"'.ord, '\\'.ord
-        io << '\\'
-      when 0x00..0x1F, 0x7F
-        raise ArgumentError.new("String contained invalid character #{byte.chr.inspect}")
+    string.gsub(io, ascii_only: true) do |char|
+      case char
+      when '\t', ' ', '"', '\\'
+        next '\\', char
+      when '\u{00}'..'\u{1F}', '\u{7F}'
+        raise ArgumentError.new("String contained invalid character #{char.inspect}")
       end
-      io.write_byte byte
     end
   end