@ysbaddaden 2018-07-17
I don't understand your benchmarks: new is always slower than old ?!