@RX14 2018-07-17
if old is always better than new in the non-ascii-only case, then why not keep the old implementation in that case?