@straight-shoota 2018-07-17
I'm not sure why the non-ascii performance is so bad. In my initial benchmarks, it showed significant improvement. And the implementation is basically the same as for ascii-only. This needs investigation.