@r00ster91 2019-11-05
@straight-shoota in the past I've also had benchmarking issues, but sometimes it actually helped to clear the cache: `sudo rm -rf ~/.cache/crystal/`. It can be glitchy sometimes.