@asterite 2019-11-13
Yeah, there's a problem with how the compiler's cache works, I think in release it doesn't work quite well. I'll have to investigate, but I'll need reproducible code(s).